+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- in `EvmTxsFetcher` fix case when we get error on first block and we checkpoint all blocks
 
 ## [5.12.1] - 2024-05-20
 ### Added